requirements.txt
django==4.1.12
mysqlclient==2.1.1
django-urls
djangorestframework
djangorestframework-simplejwt
django-cors-headers